+++ openpifpaf/functional.pyx
@@ -6,7 +6,7 @@ import numpy as np
 
 @cython.boundscheck(False)
 @cython.wraparound(False)
-def scalar_square_add_constant(double[:, :] field, x_np, y_np, width_np, double[:] v):
+def scalar_square_add_constant(float[:, :] field, x_np, y_np, width_np, float[:] v):
     minx_np = np.round(x_np - width_np).astype(np.int)
     minx_np = np.clip(minx_np, 0, field.shape[1] - 1)
     miny_np = np.round(y_np - width_np).astype(np.int)
@@ -31,7 +31,7 @@ def scalar_square_add_constant(double[:, :] field, x_np, y_np, width_np, double[
 @cython.boundscheck(False)
 @cython.wraparound(False)
 @cython.cdivision(True)
-def cumulative_average(double[:, :] ca, double[:, :] cw, x_np, y_np, width_np, double[:] v, double[:] w):
+def cumulative_average(float[:, :] ca, float[:, :] cw, x_np, y_np, width_np, float[:] v, float[:] w):
     minx_np = np.round(x_np - width_np).astype(np.int)
     minx_np = np.clip(minx_np, 0, ca.shape[1] - 1)
     miny_np = np.round(y_np - width_np).astype(np.int)
@@ -59,7 +59,7 @@ def cumulative_average(double[:, :] ca, double[:, :] cw, x_np, y_np, width_np, d
 @cython.boundscheck(False)
 @cython.wraparound(False)
 @cython.cdivision(True)
-def scalar_square_add_gauss(double[:, :] field, x_np, y_np, sigma_np, v_np, double truncate=2.0):
+def scalar_square_add_gauss(float[:, :] field, x_np, y_np, sigma_np, v_np, float truncate=2.0):
     sigma_np = np.maximum(1.0, sigma_np)
     width_np = np.maximum(1.0, truncate * sigma_np)
     minx_np = np.round(x_np - width_np).astype(np.int)
@@ -71,19 +71,19 @@ def scalar_square_add_gauss(double[:, :] field, x_np, y_np, sigma_np, v_np, doub
     maxy_np = np.round(y_np + width_np).astype(np.int)
     maxy_np = np.clip(maxy_np + 1, miny_np + 1, field.shape[0])
 
-    cdef double[:] x = x_np
-    cdef double[:] y = y_np
-    cdef double[:] sigma = sigma_np
+    cdef float[:] x = x_np
+    cdef float[:] y = y_np
+    cdef float[:] sigma = sigma_np
     cdef long[:] minx = minx_np
     cdef long[:] miny = miny_np
     cdef long[:] maxx = maxx_np
     cdef long[:] maxy = maxy_np
-    cdef double[:] v = v_np
+    cdef float[:] v = v_np
 
     cdef Py_ssize_t i, xx, yy
     cdef Py_ssize_t l = minx.shape[0]
-    cdef double deltax, deltay
-    cdef double vv
+    cdef float deltax, deltay
+    cdef float vv
 
     for i in range(l):
         for xx in range(minx[i], maxx[i]):
@@ -97,23 +97,23 @@ def scalar_square_add_gauss(double[:, :] field, x_np, y_np, sigma_np, v_np, doub
 @cython.boundscheck(False)
 @cython.wraparound(False)
 @cython.cdivision(True)
-def weiszfeld_nd(x_np, y_np, double[:] weights=None, double epsilon=1e-8, Py_ssize_t max_steps=20):
+def weiszfeld_nd(x_np, y_np, float[:] weights=None, float epsilon=1e-8, Py_ssize_t max_steps=20):
     """Weighted Weiszfeld algorithm."""
     if weights is None:
         weights = np.ones(x_np.shape[0])
 
-    cdef double[:, :] x = x_np
-    cdef double[:] y = y_np
-    cdef double[:, :] weights_x = np.zeros_like(x)
+    cdef float[:, :] x = x_np
+    cdef float[:] y = y_np
+    cdef float[:, :] weights_x = np.zeros_like(x)
     for i in range(weights_x.shape[0]):
         for j in range(weights_x.shape[1]):
             weights_x[i, j] = weights[i] * x[i, j]
 
-    cdef double[:] prev_y = np.zeros_like(y)
-    cdef double[:] y_top = np.zeros_like(y)
-    cdef double y_bottom
+    cdef float[:] prev_y = np.zeros_like(y)
+    cdef float[:] y_top = np.zeros_like(y)
+    cdef float y_bottom
     denom_np = np.zeros_like(weights)
-    cdef double[:] denom = denom_np
+    cdef float[:] denom = denom_np
 
     for s in range(max_steps):
         prev_y[:] = y
@@ -138,7 +138,7 @@ def weiszfeld_nd(x_np, y_np, double[:] weights=None, double epsilon=1e-8, Py_ssi
 
 @cython.boundscheck(False)
 @cython.wraparound(False)
-def paf_mask_center(double[:, :] paf_field, double x, double y, double sigma=1.0):
+def paf_mask_center(float[:, :] paf_field, float x, float y, float sigma=1.0):
     mask_np = np.zeros((paf_field.shape[1],), dtype=np.uint8)
     cdef unsigned char[:] mask = mask_np
 
@@ -155,9 +155,9 @@ def paf_mask_center(double[:, :] paf_field, double x, double y, double sigma=1.0
 
 @cython.boundscheck(False)
 @cython.wraparound(False)
-def paf_center(double[:, :] paf_field, double x, double y, double sigma=1.0):
+def paf_center(float[:, :] paf_field, float x, float y, float sigma=1.0):
     result_np = np.empty_like(paf_field)
-    cdef double[:, :] result = result_np
+    cdef float[:, :] result = result_np
     cdef unsigned int result_i = 0
     cdef bint take
